// PAGridCustomizer/types.ts
export type CellRendererOverrides = {
  [dataType in ColumnDataType]?: (props: CellRendererProps, rendererParams: GetRendererParams)
    => React.ReactElement | null | undefined;
};
// ...
export interface CellRendererProps {

  /** Raw value of the cell */
  value: unknown;
  /** True when grid is in RTL mode */
  isRTLMode?: boolean;
  /** Callback indicating the grid cell has been clicked */
  onCellClicked?: (event?: React.MouseEvent<HTMLElement, MouseEvent> | MouseEvent) => void;
  /** Returns the validation error for the cell */
  validationError?: Error | null;
  /** True when cell should be right aligned */
  isRightAligned?: boolean;
  /** Callback to pragmatically start editing the cell */
  startEditing?: (editorInitValue?: unknown) => void;
  /** True when cell is on the last row of the grid */
  isLastRow?: boolean;
  /** Grid row height in pixels */
  rowHeight?: number;
  /** Cell column data type */
  columnDataType?: ColumnDataType;
  /** Formatted value of the cell */
  formattedValue?: string;
  /** HTML element containing the cell */
  cellContainerElement?: HTMLElement;
  /** Cell error label Id */
  cellErrorLabelId?: string;
  /** True when the cell column is editable */
  columnEditable?: boolean;
}

export interface GetRendererParams {

  /** Column definitions for all visible columns in the grid */
  colDefs: ColumnDefinition[];
  /** Cell column index */
  columnIndex: number;
  /** Cell row data */
  rowData?: RowData;
  /** True when tab navigation is allowed in the grid */
  allowTabKeyNavigation?: boolean;
  setExpanded?: 1;
  getExpanded():unknown

}
// ...
export const RECID = '__rec_id';
export const RSTAT = '__row_status';
export const AGROW = 'ag-row';
export const GRCOL = '__group_column';
export const GRCNT = '__group_count';

export interface RowData {
  /** Unique id for the row */
  [RECID]: string;
  [RSTAT]: string;
  [AGROW]: string;
  [GRCOL]: string;
  [GRCNT]: string;
}

export interface ColumnDefinition {
  /** Field name (should be unique) */
  name: string;
  /** Column display name. Will be shown as column header.If none provided shimmer will be shown in place of header */
  displayName?: string;
  /** The unique ID to give the column. This is optional. If missing, the ID will default to the name.
   * If defined, this ID will be used to identify the column in the API for sorting and filtering. */
  colId?: string;
  /** Data type for the values in the column. */
  dataType: string;
  /** Custom Renderer type to be used for this column when custom renderer override is provided. */
  CustomRendererType?: string;
  /** Width of the column. Auto calculated if not set. */
  width?: number;
  /** Min width, in pixels, of the cell */
  minWidth?: number;
  /** Max width, in pixels, of the cell */
  maxWidth?: number;
  /** True if column is initially hidden
   * @Notice To show/hide columns after the first render use PAGridAPI.setColumnVisible()
   */
  hide?: boolean;
  /** Column is primary field */
  isPrimary: boolean;
  getFormattedValue : (id ?: string) => string; //UNSUPPORTED
}

export type ColumnDataType =
  | 'Text'
  | 'Email'
  | 'Phone'
  | 'Ticker'
  | 'URL'
  | 'TextArea'
  | 'Lookup'
  | 'Customer'
  | 'Owner'
  | 'MultiSelectPicklist'
  | 'OptionSet'
  | 'TwoOptions'
  | 'Duration'
  | 'Language'
  | 'Multiple'
  | 'TimeZone'
  | 'Integer'
  | 'Currency'
  | 'Decimal'
  | 'FloatingPoint'
  | 'AutoNumber'
  | 'DateOnly'
  | 'DateAndTime'
  | 'Image'
  | 'File'
  | 'Persona';